Fix the damn thing that was causing lots of unnecessary recompositions tanking the performance.
The problem was in code that is calling postsScreenViewModel.onPostListBuilt(), namely we were
re-executing the checks to figure out whether we have a visible PostCell on every recomposition
which, for some reason, was causing more recompositions to happen (basically infinite). I have
no idea why it's causing more recompositions because nothing in LazyListState is modified.

+++ app/src/main/java/com/github/k1rakishou/kurobaexlite/ui/helpers/Modifiers.kt
@@ -24,8 +24,8 @@ fun Modifier.simpleVerticalScrollbar(
   state: LazyListState,
   chanTheme: ChanTheme,
   contentPadding: PaddingValues = DefaultPaddingValues,
-  scrollbarWidth: Float,
-  scrollbarMinHeight: Float,
+  scrollbarWidth: Int,
+  scrollbarMinHeight: Int,
   scrollbarDragged: Boolean
 ): Modifier {
   return composed(
@@ -77,7 +77,7 @@ fun Modifier.simpleVerticalScrollbar(
             bottomPaddingPx = bottomPaddingPx,
             layoutInfo = layoutInfo,
             firstVisibleElementIndex = firstVisibleElementIndex,
-            scrollbarMinHeight = scrollbarMinHeight,
+            scrollbarMinHeight = scrollbarMinHeight.toFloat(),
             realScrollbarHeightDiff = null
           )
 
@@ -87,7 +87,7 @@ fun Modifier.simpleVerticalScrollbar(
           drawRect(
             color = chanTheme.textColorHintCompose,
             topLeft = Offset(offsetX, offsetY),
-            size = Size(scrollbarWidth, scrollbarHeightAdjusted),
+            size = Size(scrollbarWidth.toFloat(), scrollbarHeightAdjusted),
             alpha = alpha
           )
         }
